backend/requirements.txt

fastapi
uvicorn[standard]
neo4j
asyncpg
openai
faiss-cpu
python-dotenv
pydantic
